First version of a write version command
A simple assignment-based replacer which can handle quoted strings in
either quote style.

This isn't regex-based, but we might have a future regex-based
version, so the version writing characteristics here are called
"assign" -- meaning it's looking for an assignment and using that.

This also introduces some 'pyproject.toml'-backed config for the
version writing mode. The default used as a fallback is to look for
`version = ...` in `pyproject.toml`.