Enhance support for extras, split wheel data reader
Move nontrivial wheel metadata interpretation into a dedicated module,
and enhance it with the ability to strip an extra marker specification
back out of the dist dependency fields. It does this with simplistic
regex matches.

This is also now used to refine the non-extra dependencies listing when
pulled from wheel metadata, to remove the dependencies which appear to be
associated with a Provides-Extra field.