Update extra parsing to produce originals as well
Do all of the parsing work on optional dependencies at once. The regex
matches which "find matching requirements" can also be used to do the
removal of "the extra parts". This makes the handling more uniform for
any usages which will access both parts of the metadata.

The original extras are now exposed, with the "cleaned" ones not yet
provided in the API.