Add a CLI command to read optional-dependencies
Also fix issues with name normalization which were revealed by this
testing. The extra names produced are now always normalized, even when
reading static metadata. Conflicts due to improper static metadata are
raised as errors.

The CLI command exposes the option of preserving exact wheel metadata,
with an explanation in the help text.